requirements.txt
# FastAPI service
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
numpy>=1.24.0
opencv-python-headless>=4.8.0

# Dev / test
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiohttp>=3.9.0
